+++ src/melclouddevice.js
@@ -169,9 +169,10 @@ class MELCLOUDDEVICE extends EventEmitter {
                     const sceneOwner = deviceState.SceneOwner;
 
                     const melCloudInfo = this.melCloudInfo;
-                    const useFahrenheit = melCloudInfo.UseFahrenheit ? 1 : 0;
+                    const useFahrenheit = (melCloudInfo.UseFahrenheit == true) ? 1 : 0;
+                    const lockPhysicalControls = (prohibitSetTemperature == true || prohibitOperationMode == true || prohibitPower == true) ? 1 : 0;
                     this.emit('deviceInfo', manufacturer, modelName, modelName1, serialNumber, firmwareRevision);
-                    this.emit('deviceState', melCloudInfo, useFahrenheit, deviceState, power, inStandbyMode, operationMode, roomTemperature, setTemperature, setFanSpeed, numberOfFanSpeeds, vaneHorizontal, vaneVertical);
+                    this.emit('deviceState', melCloudInfo, useFahrenheit, deviceState, power, inStandbyMode, operationMode, roomTemperature, setTemperature, setFanSpeed, numberOfFanSpeeds, vaneHorizontal, vaneVertical, lockPhysicalControls);
                 } catch (error) {
                     this.emit('error', `Check device state error: ${error}`);
                 };